refactor(runtime-sdk): разнести sdk-host god-file на host/ + providers/
Механический, поведение-сохраняющий вынос (#6 из sdk-runtime-review).
sdk-host.ts (871 строк) → тонкий barrel + entry-point:

  host/session-host.ts   — SessionHost core (юнит-тест без сокетов/SDK)
  host/socket-server.ts  — net server + backpressure + control-команды + dispatch + readAppendSystemPrompt
  host/history-log.ts    — durable events.ndjson sink (createEventLogSink)
  providers/claude-agent-sdk.ts — default Claude путь (runClaudeAgentMode), MiMo full-agent шарит его
  providers/openai-compatible.ts — GLM/MiMo legacy chat-loop (runOpenAiCompatMode/runGlmMode)
  providers/mimo-anthropic.ts    — MiMo Anthropic-endpoint env (applyMimoAnthropicEnv)

sdk-host.ts ОСТАЁТСЯ entry (index.ts резолвит sdk-host.js) + re-export
barrel (тесты импортят SessionHost/SessionHostOptions/backpressure из
../sdk-host.js без правок). Атомарный коммит: split компилится только
целиком. Готовит почву под Codex/OpenAI-provider (Фаза 2).

Проверки: build зелёный, тесты 51/2 (без изменений), standalone smoke
(READY+status+kill exit0).
